+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantHandler.cs
@@ -3,11 +3,14 @@ using Microsoft.Extensions.Logging;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
 using Restaurants.Domain.Repositories;
+using Restaurants.Domain.Interfaces;
+using Restaurants.Domain.Constants;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
 
 internal class DeleteRestaurantHandler(ILogger<DeleteRestaurantHandler> logger,
-    IRestaurantsRepository restaurantsRepository) 
+    IRestaurantsRepository restaurantsRepository,
+    IRestaurantAuthorizationService restaurantAuthorization) 
     : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
@@ -16,7 +19,9 @@ internal class DeleteRestaurantHandler(ILogger<DeleteRestaurantHandler> logger,
 
         var restaurant = await restaurantsRepository.GetSpecificAsync(request.Id)
             ?? throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
-        
+
+        if (!restaurantAuthorization.Authorize(restaurant, ResourceOperations.Delete))
+            throw new ForbidException();
 
         await restaurantsRepository.Delete(restaurant);
